[Word2007Addin] Word validator deprecation
For office 2010 and superior, it is recommended by the DAISY consortium
to use the Microsoft Accessibility checker to validate the word structure.
The validate button now acts as a shortcut for the accessibility checker on
office 2010 and newer.
This tools is allegedly only available starting office 2010, so the validator
remain used for office 2003 and 2007.