[XSLT] Port Daisyclass.cs to Java, move oox2Daisy.xsl to XSLT 2.0 and wrap in XProc step
and run it with XMLCalabash/Saxon.

The C# code was ported "as is" as much as possible, including bugs and
bad code, to keep things uncluttered. Optimization will be done
later. Some variable types were changed to match the types of
corresponding XSLT variables. Unused C# code was dropped.

To make the whole transformation contained within the XSLT, the
DaisyClass object is now constructed by the XSLT itself and the
DaisyClass arguments are defined as XSLT parameters.

There are a number of backward incompatibilities in XSLT 2.0 that
needed to be addressed. The most important issue is that XSLT 1.0 is
less typed and does more automatic casting. XSLT 1.0 is also less
strict when it comes to providing parameters to templates. There is
also a difference in the behavior of the (sometimes implicit) string()
function.

Note that the tests don't work yet.