[XSLT] Remove progress messages
The way the progress indication worked made very little sense: first
the total duration of the conversion was computed by actually running
the conversion and counting the total number of progress
messages. Then, the conversion was run a second time, indicating the
progress as a fraction of the total duration.