[Refactoring] Code rewrite and cleanup for 2.6.1.3
The 2 main libraries have been reorganized:
- The DaisyConverterLib now handles the conversion processes and settings,
and it provides interfaces for preprocessing and event or message handling.
This library is moved to the Daisy.SaveAsDAISY.Conversion namespace.
- The DaisyAddinLib now focus on the UI elements and associated classes
and resources that are to be used by addin.
This library is moved to the Daisy.SaveAsDAISY.Forms namespace.

Addins and Command line tool is using the updated library.
Addins project and classes have been moved to Daisy.SaveAsDAISY.Addin
namespace, with Word2007 and Word2003 subnamespaces for the respective
addin.
Each addins are implementing the preprocessing interface.

The following projects have been removed :
- WordXP addin (end of support)
- Addin project (Merged during refactoring with the UI library)
- Converter project (Merged during refactoring with the conversion library)

Conversion library updates:
- Forms, dialogs and controls are moved to the UI library
- Conversion logic from preprocessing to postprocessing is now
completely handled by the converter class, using the preprocessDocument and
convert methods. The convert method can take either one document or a
list of document (to be merged in one output)
- Messages and event handling is delegated using the
IConversionEventHandler interface.
- Preprocessing is delegated using a new IDocumentPreprocessor
interface.
- Conversion settings are dispatch in 3 main class :
  - ConverterSettings handle the global settings for all conversions
  - ConversionParameters handle the settings for the conversion of one
  or more document
  - DocumentParameters handle the settings and preprocessing data of a
  single document. This class also handle the list of subdocuments
  parameters.
- Conversion of a document to DTBook XML is handled by the
WordToDTBookXMLTransform class using the ConvertDocument method.
Conversion methods have been merged in a single ConvertDocument method
that uses a DocumentParameter and a ConversionParameters arguments. An
optional XmlDocument can be added to merge the conversion result with a
previous conversion.

UI library:
- Conversion logic related classes are moved to the conversion library
- Conversion progress bars are replaced by a message dialog linked to
an implementation of the event handler interface.
- The UI Library include an overload of the Converter class to manage
the message dialog and implements user interactions for requests (files
selections and conversion parameters).

A preliminary cleanup has been applied on the global code base.