feat(build): new powershell build script
Starting the preparation of a simple build script to simplify the creation of new
installers.
A `refreshpipeline` switch allow the script to update the provided DAISY Pipeline 2 by reparsing its directory to update the wix project.
A version parameter is under development to also simplify the update and the generation of an installer with a new version number.